refactor: adjust project_api_key to new file structure
