+++ internal/service/projectapikey/data_source_project_api_keys.go
@@ -1,4 +1,4 @@
-package mongodbatlas
+package projectapikey
 
 import (
 	"context"
