+++ internal/service/projectapikey/resource_project_api_key.go
@@ -1,4 +1,4 @@
-package mongodbatlas
+package projectapikey
 
 import (
 	"context"
