+++ mongodbatlas/data_source_org_id.go
@@ -8,6 +8,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/id"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/mongodb/terraform-provider-mongodbatlas/internal/config"
+	"github.com/mongodb/terraform-provider-mongodbatlas/internal/service/projectapikey"
 
 	matlas "go.mongodb.org/atlas/mongodbatlas"
 )
@@ -42,7 +43,7 @@ func dataSourceMongoDBAtlasOrgIDRead(ctx context.Context, d *schema.ResourceData
 	for idx, role := range apiKeyOrgList.APIKey.Roles {
 		if strings.HasPrefix(role.RoleName, "ORG_") {
 			if err := d.Set("org_id", apiKeyOrgList.APIKey.Roles[idx].OrgID); err != nil {
-				return diag.Errorf(ErrorProjectSetting, `org_id`, root.APIKey.ID, err)
+				return diag.Errorf(projectapikey.ErrorProjectSetting, `org_id`, root.APIKey.ID, err)
 			}
 			d.SetId(apiKeyOrgList.APIKey.Roles[idx].OrgID)
 			return nil
